+++ l2rpn_baselines/test/test_import.py
@@ -70,14 +70,15 @@ class TestPandapowerOPFAgent(TestImport, unittest.TestCase):
         return "PandapowerOPFAgent"
 
 
-class TestPandapowerGeirina(TestImport, unittest.TestCase):
-    def load_module(self):
-        return "Geirina"
+# because it deactivates the eager mode
+# class TestPandapowerGeirina(TestImport, unittest.TestCase):
+#     def load_module(self):
+#         return "Geirina"
 
 
-class TestAsynchronousActorCritic(TestImport, unittest.TestCase):
-    def load_module(self):
-        return "AsynchronousActorCritic"
+# class TestAsynchronousActorCritic(TestImport, unittest.TestCase):
+#     def load_module(self):
+#         return "AsynchronousActorCritic"
 
 
 if __name__ == "__main__":